fix: 2.5.0 1. 修复postgres兼容问题

+++ easy-retry-datasource/easy-retry-datasource-template/src/main/java/com/aizuda/easy/retry/template/datasource/persistence/po/Namespace.java
@@ -52,5 +52,5 @@ public class Namespace implements Serializable {
     /**
      * 逻辑删除 1、删除
      */
-    private Byte deleted;
+    private Integer deleted;
 }
